[refactor] data 파일 패키징
데이터 파일 패키징
