+++ app/src/main/java/meaning/morning/Presentation/PictureRecyclerviewFragment.kt
@@ -2,12 +2,12 @@ package meaning.morning.presentation
 
 import android.content.Context
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
+import meaning.morning.Presentation.data.MyFeedPictureData
 import meaning.morning.R
 import meaning.morning.databinding.FragmentPictureRecyclerviewBinding
 import meaning.morning.presentation.adapter.MyFeedPictureAdapter
