+++ app/src/main/java/meaning/morning/Presentation/data/MyFeedPictureData.kt
@@ -1,4 +1,4 @@
-package meaning.morning.presentation
+package meaning.morning.Presentation.data
 
 import android.widget.ImageView
