+++ app/src/main/res/layout/item_list.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="pictureList"
-            type="meaning.morning.presentation.MyFeedPictureData"/>
+            type="meaning.morning.Presentation.data.MyFeedPictureData"/>
 
     </data>
